Testing broken GLOW daytime

+++ src/glowpython2/igrf.py
@@ -18,7 +18,7 @@ class Igrf(Singleton):
         # Initialize IGRF model parameters
         glowigrf.igrf_init(str(DATADIR), logfile)
     
-    def dipangle(self, time: Numeric, lat: Numeric, lon: Numeric, alt: Numeric | np.ndarray) -> Numeric | np.ndarray:
+    def dipangle(self, time: Numeric, lat: Numeric, lon: Numeric, alt: Numeric | np.ndarray) -> float | np.ndarray:
         """Calculate magnetic dip angle in degrees.
 
         Args:
